言及の処理をモジュールに変更
