+++ app/models/report.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class Report < ApplicationRecord
-  include ReportsHelper
+  include Mentionable
   belongs_to :user
   has_many :comments, as: :commentable, dependent: :destroy
